x86_64: removed util.c and redistributed it's functions
